daddy-joke-telegram-bot/requirements.txt
python-telegram-bot==8.1.1
emoji==0.5.1
requests==2.20.1
